Use 'MINIO_INST_COUNT' in SAPI as the number of instances we need
This removes MANTA_MINIO_MIN_INST and MANTA_MINIO_MAX_INST. The
former didn't really make sense, as even though we could set a lower
bound of the number of instances that need to be online, the server
wouldn't actually start until all instances are at least registered in
the nameservice, checked for by the start-method script.

Whether the minio service is functioning on all of those instances
is a different matter.